requirements.txt
flask>=2.0
python-dotenv>=0.21
python-telegram-bot>=13.15
twilio>=8.0
scikit-learn>=1.0
numpy>=1.22
requests>=2.28
pytest>=7.0
PyYAML>=6.0

